requirements.txt
Flask
Flask-Cors
torch
torchvision
transformers
Pillow
accelerate
requests
sentence-transformers
faiss-cpu

#source .venv/bin/activate 
#pip install -r requirements.txt
